@@ -1,18 +1,21 @@
-import adapter from "@sveltejs/adapter-static";
-import preprocess from 'svelte-preprocess';
-import { mdsvex } from "mdsvex";
+import { mdsvex } from 'mdsvex';
+import adapter from '@sveltejs/adapter-static';
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-    extensions: [".svelte", ".svelte.md"],
-    preprocess: [mdsvex({
-            extensions: [".svelte.md"],
-        }),
-        preprocess()
-    ],
+    // Consult https://svelte.dev/docs/kit/integrations
+    // for more information about preprocessors
+    preprocess: [vitePreprocess(), mdsvex()],
+
     kit: {
-        adapter: adapter(),
+        // adapter-auto only supports some environments, see https://svelte.dev/docs/kit/adapter-auto for a list.
+        // If your environment is not supported, or you settled on a specific environment, switch out the adapter.
+        // See https://svelte.dev/docs/kit/adapters for more information about adapters.
+        adapter: adapter()
     },
-    };
+
+    extensions: ['.svelte', '.svx']
+};
 
 export default config;
